Added comments via RabbitMQ

+++ src/main/java/net/milanvit/learningspringboot/InitDatabase.java
@@ -1,5 +1,6 @@
 package net.milanvit.learningspringboot;
 
+import net.milanvit.learningspringboot.images.Image;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.data.mongodb.core.MongoOperations;
@@ -10,15 +11,17 @@ public class InitDatabase {
     @Bean
     CommandLineRunner init(MongoOperations operations) {
         return args -> {
-            operations.dropCollection(Image.class);
+            if (false) {
+                operations.dropCollection(Image.class);
 
-            operations.insert(new Image("1", "learning-spring-boot-cover.jpg"));
-            operations.insert(new Image("2", "learning-spring-boot-2nd-edition-cover.jpg"));
-            operations.insert(new Image("3", "bazinga.png"));
+                operations.insert(new Image("1", "learning-spring-boot-cover.jpg"));
+                operations.insert(new Image("2", "learning-spring-boot-2nd-edition-cover.jpg"));
+                operations.insert(new Image("3", "bazinga.png"));
 
-            operations.findAll(Image.class).forEach(image -> {
-                System.out.println(image.toString());
-            });
+                operations.findAll(Image.class).forEach(image -> {
+                    System.out.println(image.toString());
+                });
+            }
         };
     }
 }
